[CAS-61] Pull up common implementation from ChatParser into default methods
